+++ src/Frontend/Dfe.Complete/Pages/Projects/TaskList/Tasks/ReceiveGrantPaymentCertificateTask/ReceiveGrantPaymentCertificateTask.cshtml.cs
@@ -8,10 +8,10 @@ using Dfe.Complete.Services;
 using MediatR;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
-namespace Dfe.Complete.Pages.Projects.TaskList.Tasks.DeclarationOfExpenditureCertificateTask
+namespace Dfe.Complete.Pages.Projects.TaskList.Tasks.ReceiveGrantPaymentCertificateTask
 {
-    public class DeclarationOfExpenditureCertificateTaskModel(ISender sender, IAuthorizationService authorizationService, ILogger<DeclarationOfExpenditureCertificateTaskModel> logger, ErrorService errorService)
-    : BaseProjectTaskModel(sender, authorizationService, logger, NoteTaskIdentifier.DeclarationOfExpenditureCertificate)
+    public class ReceiveGrantPaymentCertificateTaskModel(ISender sender, IAuthorizationService authorizationService, ILogger<ReceiveGrantPaymentCertificateTaskModel> logger, ErrorService errorService)
+    : BaseProjectTaskModel(sender, authorizationService, logger, NoteTaskIdentifier.ReceiveGrantPaymentCertificate)
     {
         [BindProperty(Name = "not-applicable")]
         public bool? NotApplicable { get; set; }
@@ -33,10 +33,11 @@ namespace Dfe.Complete.Pages.Projects.TaskList.Tasks.DeclarationOfExpenditureCer
             await base.OnGetAsync();
             Type = Project.Type;
             TasksDataId = Project.TasksDataId?.Value;
-            NotApplicable = TransferTaskData.DeclarationOfExpenditureCertificateNotApplicable;
-            CheckCertificate = TransferTaskData.DeclarationOfExpenditureCertificateCorrect;
-            ReceivedDate = TransferTaskData.DeclarationOfExpenditureCertificateDateReceived;
-            Saved = TransferTaskData.DeclarationOfExpenditureCertificateSaved;
+
+            NotApplicable = ConversionTaskData.ReceiveGrantPaymentCertificateNotApplicable;
+            CheckCertificate = ConversionTaskData.ReceiveGrantPaymentCertificateCheckCertificate;
+            ReceivedDate = ConversionTaskData.ReceiveGrantPaymentCertificateDateReceived;
+            Saved = ConversionTaskData.ReceiveGrantPaymentCertificateSaveCertificate;
 
             return Page();
         }