+++ CHANGELOG.md
@@ -18,9 +18,13 @@ Security in case of vulnerabilities.
 
 ### Added
 
+- Added `Confirm all condition has been met` task for conversion projects.
+
 ### Changed
 
 ### Fixed
+
+- Fixed broken task note urls.
 
 See the [full commit history](https://github.com/DFE-Digital/complete-conversions-transfers-changes/compare/production-2025-09-25.771...main) for everything awaiting release
 
@@ -29,9 +33,11 @@ See the [full commit history](https://github.com/DFE-Digital/complete-conversion
 ## [1.18.0](https://github.com/DFE-Digital/complete-conversions-transfers-changes/releases/tag/production-2025-09-25.771) - 2025-09-25
 
 ### Removed
+
 - removed the 'Add Project' in your project button and disabled access to all related subpages 
 
 ### Added
+
 - Ability to edit project information for conversions and transfers
 - ProjectGroup is created when a project is edited with a new GRN
 - Added Add, Edit and Delete External Contacts for both conversion and transfer projects.